// frontend/src/domain/invoiceCustomFields/invoiceCustomFields.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type InvoiceCustomFieldsVariables = {
  scope: string;
};

export type InvoiceCustomFieldsResult = {
  customFields: ({
  totalCount: number;
  nodes: Array<{
  id: string;
  key: string;
  name: string;
  valueType: "INTEGER" | "TEXT" | "DATE" | "REAL" | "OPTION" | "BOOLEAN";
  displayMode: "HIDDEN" | "VISIBLE" | "PROMINENT" | "OTHER" | null;
  options: Array<{
  id: string;
  key: string;
  name: string;
  parentOptionId: string | null;
}>;
}>;
});
};

export const InvoiceCustomFields = {
  query: "query invoiceCustomFields($scope: String!) {\n  customFields(filter: {scope: {equalTo: $scope}}) {\n    ... on CustomFieldConnector {\n      totalCount\n      nodes {\n        id\n        key\n        name\n        valueType\n        displayMode\n        options {\n          id\n          key\n          name\n          parentOptionId\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<InvoiceCustomFieldsResult, InvoiceCustomFieldsVariables>;